ADPL_master/requirements.txt
pytorch==1.7.1
torchvision==0.8.2
imageio==2.9.0
kornia==0.2.0
scipy==1.2.1
tensorboard==2.3.0
tqdm==4.54.1
scipy==1.2.1
opencv-python==3.4.2
